requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
msgspec==0.18.6
numpy==2.0.0
openai==1.35.9
pandas==2.2.2
pydantic==2.8.0
pydantic_core==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
